Stock %change Predictor over 5 days
Posting this today 02-16-2023 and will update on 02-19-2023 after closing to see if the apple stock changed as predicted. All of the graphs are sloppy and some do not make much sense. I am working on cleaning this up as well as adding tables to see if this is the better option to display the data. I will get back to this in 5 day to see if the prediction is correct!!